--- EqualityInformationApi.Tests/AwsMockWebApplicationFactory.cs
@@ -1,71 +1,0 @@
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.DynamoDBv2.Model;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using System;
-using System.Collections.Generic;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Hackney.Core.DynamoDb;
-
-namespace EqualityInformationApi.Tests
-{
-    public class AwsMockWebApplicationFactory<TStartup>
-            : WebApplicationFactory<TStartup> where TStartup : class
-    {
-        private readonly List<TableDef> _tables;
-
-        public IAmazonDynamoDB DynamoDb { get; private set; }
-        public IDynamoDBContext DynamoDbContext { get; private set; }
-        public AwsMockWebApplicationFactory(List<TableDef> tables)
-        {
-            _tables = tables;
-        }
-
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.ConfigureAppConfiguration(b => b.AddEnvironmentVariables())
-                .UseStartup<Startup>();
-            builder.ConfigureServices(services =>
-            {
-                var url = Environment.GetEnvironmentVariable("DynamoDb_LocalServiceUrl");
-                var snsUrl = Environment.GetEnvironmentVariable("Localstack_SnsServiceUrl");
-
-                services.AddSingleton<IAmazonDynamoDB>(sp =>
-                {
-                    var clientConfig = new AmazonDynamoDBConfig { ServiceURL = url };
-                    return new AmazonDynamoDBClient(clientConfig);
-                });
-
-
-                services.ConfigureSns();
-
-                var serviceProvider = services.BuildServiceProvider();
-                DynamoDb = serviceProvider.GetRequiredService<IAmazonDynamoDB>();
-                DynamoDbContext = serviceProvider.GetRequiredService<IDynamoDBContext>();
-
-                EnsureTablesExist(DynamoDb, _tables);
-            });
-        }
-
-        private static void EnsureTablesExist(IAmazonDynamoDB dynamoDb, List<TableDef> tables)
-        {
-            foreach (var table in tables)
-            {
-                try
-                {
-                    var request = new CreateTableRequest(table.Name,
-                        new List<KeySchemaElement> { new KeySchemaElement(table.KeyName, KeyType.HASH) },
-                        new List<AttributeDefinition> { new AttributeDefinition(table.KeyName, table.KeyType) },
-                        new ProvisionedThroughput(3, 3));
-                    _ = dynamoDb.CreateTableAsync(request).GetAwaiter().GetResult();
-                }
-                catch (ResourceInUseException)
-                {
-                    // It already exists :-)
-                }
-            }
-        }
-    }
-}
